Workable version, totally
still some issues with aspect ratio. 
seems it can't find the aspect ratio i find. i want 9:16 but with wide lens. 
it only gives me 4:3 with wide lens, which then i save as 3:4 i think. aspect ratio is good and it's wide, but when on a 9:16 timeline, it loses some stuff on the side. to be improved